+++ src/providers/Preview.ts
@@ -1,33 +1,74 @@
 import * as vscode from "vscode";
 import * as path from 'path';
 import * as fs from "fs";
-import { parsedDocuments, diagnostics, diagnosticCollection, getEditor, activePreviewChanged, parseDocument } from "../extension";
+import { parsedDocuments, diagnostics, diagnosticCollection, getEditor, parseDocument } from "../extension";
 import { getFountainConfig } from "../configloader";
 import { TopmostLineMonitor } from "../utils/topMostLineMonitor";
 
-export var previewpanels = new Map<string, vscode.WebviewPanel>();
+interface preview{
+    uri:string;
+    dynamic:boolean;
+    panel:vscode.WebviewPanel;
+    id:Number;
+}
+
+export var previews:preview[] = [];
 var isScrolling = false;
 
-export function createPreviewPanel(editor:vscode.TextEditor): vscode.WebviewPanel{
+export function getPreviewPanels(docuri:vscode.Uri):preview[]{
+    let selectedPreviews:preview[] = []
+    for (let i = 0; i < previews.length; i++) {
+        if(previews[i].uri == docuri.toString())
+            selectedPreviews.push(previews[i])
+    }
+    return selectedPreviews;
+}
+export function removePreviewPanel(id:Number){
+    for (let i = 0; i < previews.length; i++) {
+        if(previews[i].id == id){
+            previews.slice(i,1);
+            break;
+        }
+
+    }
+}
+
+export function getPreviewsToUpdate(docuri:vscode.Uri):preview[]{
+    let selectedPreviews:preview[] = [];
+    for (let i = 0; i < previews.length; i++) {
+        if(previews[i].uri == docuri.toString() || previews[i].dynamic)
+            selectedPreviews.push(previews[i]);
+    }
+    return selectedPreviews;
+}
+
+export function createPreviewPanel(editor:vscode.TextEditor, dynamic:boolean): vscode.WebviewPanel{
 	if(editor.document.languageId!="fountain"){
 		vscode.window.showErrorMessage("You can only preview Fountain documents as a screenplay!");
 		return undefined;
     }
     let preview:vscode.WebviewPanel;
-    if(previewpanels.has(editor.document.uri.toString())){
-        //preview already exists, simply show it
-        preview = previewpanels.get(editor.document.uri.toString());
-        preview.reveal(preview.viewColumn);
-        loadWebView(editor.document.uri, preview);
-    }
-    else{
+    let presentPreviews = getPreviewPanels(editor.document.uri);
+    presentPreviews.forEach(p => {
+        if(p.uri == editor.document.uri.toString() && p.dynamic == dynamic){
+            //The preview already exists
+            p.panel.reveal();
+            preview = p.panel;
+            dynamic = p.dynamic;
+        }
+    });
+
+    if(preview == undefined){
+        //The preview didn't already exist
+        var previewname = path.basename(editor.document.fileName).replace(".fountain","");
+        if(dynamic) previewname = "Fountain Preview";
         preview = vscode.window.createWebviewPanel(
             'fountain-preview', // Identifies the type of the webview. Used internally
-            path.basename(editor.document.fileName).replace(".fountain",""), // Title of the panel displayed to the user
+            previewname, // Title of the panel displayed to the user
             vscode.ViewColumn.Three, // Editor column to show the new webview panel in.
-            { enableScripts: true, retainContextWhenHidden:true }
-        );    
+            { enableScripts: true, retainContextWhenHidden:true });
     }
+    loadWebView(editor.document.uri, preview, dynamic);
     return preview;
 }
 
@@ -36,37 +77,31 @@ function assetsPath(): string{
     return __dirname.substr(0, __dirname.lastIndexOf(path.sep));
 }
 
-function loadWebView(docuri: vscode.Uri, preview:vscode.WebviewPanel) {
+function loadWebView(docuri: vscode.Uri, preview:vscode.WebviewPanel, dynamic:boolean) {
     var cleandir = assetsPath().split(String.fromCharCode(92)).join("/");
-    previewpanels.set(docuri.toString(), preview);
-    preview.onDidChangeViewState(e =>{
-        if(e.webviewPanel.active){
-            activePreviewChanged(docuri);
-        }
-        if(e.webviewPanel.visible){
-            console.log(previewpanels);
-        }
-    });
+    let id = Date.now()+Math.floor((Math.random()*1000));
+    previews.push({uri:docuri.toString(), dynamic:dynamic, panel:preview, id:id });
+
 	preview.webview.onDidReceiveMessage(async message => {
         if (message.command == "updateFontResult") {
-            if (message.content == false && parsedDocuments.get(docuri.toString()).properties.fontLine != -1) {
+            if (message.content == false && parsedDocuments.get(vscode.Uri.parse(message.uri).toString()).properties.fontLine != -1) {
                 //The font could not be rendered
                 diagnostics.length = 0;
-                diagnostics.push(new vscode.Diagnostic(new vscode.Range(new vscode.Position(parsedDocuments.get(docuri.toString()).properties.fontLine, 0), new vscode.Position(parsedDocuments.get(docuri.toString()).properties.fontLine, 5)), "This font could not be rendered in the live preview. Is it installed?", vscode.DiagnosticSeverity.Error));
-                diagnosticCollection.set(docuri, diagnostics);
+                diagnostics.push(new vscode.Diagnostic(new vscode.Range(new vscode.Position(parsedDocuments.get(vscode.Uri.parse(message.uri).toString()).properties.fontLine, 0), new vscode.Position(parsedDocuments.get(docuri.toString()).properties.fontLine, 5)), "This font could not be rendered in the live preview. Is it installed?", vscode.DiagnosticSeverity.Error));
+                diagnosticCollection.set(vscode.Uri.parse(message.uri), diagnostics);
             }
             else {
                 //Yay, the font has been rendered
-                diagnosticCollection.set(docuri, []);
+                diagnosticCollection.set(vscode.Uri.parse(message.uri), []);
             }
         }
         else if(message.command == "revealLine"){
-            if(!getFountainConfig(docuri).synchronized_markup_and_preview) return;
+            if(!getFountainConfig(vscode.Uri.parse(message.uri)).synchronized_markup_and_preview) return;
             isScrolling = true;
             console.log("jump to line:"+message.content);
             const sourceLine = Math.floor(message.content);
             const fraction = message.content - sourceLine;
-            let editor = getEditor(docuri);
+            let editor = getEditor(vscode.Uri.parse(message.uri));
             if(editor && !Number.isNaN(sourceLine))
             {
                 const text = editor.document.lineAt(sourceLine).text;
@@ -84,9 +119,9 @@ function loadWebView(docuri: vscode.Uri, preview:vscode.WebviewPanel) {
 
             let selectionposition = new vscode.Position(message.line, message.character);
 
-            let editor = getEditor(docuri);
+            let editor = getEditor(vscode.Uri.parse(message.uri));
             if(editor == undefined){
-                var doc = await vscode.workspace.openTextDocument(docuri);
+                var doc = await vscode.workspace.openTextDocument(vscode.Uri.parse(message.uri));
                 editor = await vscode.window.showTextDocument(doc)
             }
             else{
@@ -100,20 +135,20 @@ function loadWebView(docuri: vscode.Uri, preview:vscode.WebviewPanel) {
         }
     });
     preview.onDidDispose(()=>{
-        previewpanels.delete(docuri.toString());
+        removePreviewPanel(id);
     })
 
     preview.webview.html = webviewHtml.replace(/\$ROOTDIR\$/g, cleandir);
-    preview.webview.postMessage({ command: 'seturi', content: docuri.toString() })
+    preview.webview.postMessage({ command: 'setstate', uri: docuri.toString(), dynamic: dynamic })
     var config = getFountainConfig(docuri);
     preview.webview.postMessage({ command: 'updateconfig', content: config })
 
     var editor = getEditor(docuri);
     if(editor){
         parseDocument(editor.document);
-        preview.webview.postMessage({ command: 'highlightline', content:editor.selection.start.line})
+        if(config.synchronized_markup_and_preview)
+            preview.webview.postMessage({ command: 'highlightline', content:editor.selection.start.line})
     }
-
 }
 
 const _topmostLineMonitor = new TopmostLineMonitor();
@@ -128,33 +163,34 @@ function scrollTo(topLine: number, resource:vscode.Uri) {
     }
 
     let editor = getEditor(resource);
-    console.log(previewpanels.keys);
-	if (previewpanels.has(resource.toString()) && editor != undefined) {
-		previewpanels.get(resource.toString()).webview.postMessage({ command: 'showsourceline', content: topLine, linescount: editor.document.lineCount, source: "scroll" });
-	}
+    if(getFountainConfig(editor.document.uri).synchronized_markup_and_preview){
+        previews.forEach(p => {
+            if(p.uri == resource.toString()){
+                p.panel.webview.postMessage({ command: 'showsourceline', content: topLine, linescount: editor.document.lineCount, source: "scroll" });
+            }
+        });
+    }
+
 }
 
 
 vscode.workspace.onDidChangeConfiguration(change => {
-	previewpanels.forEach((previewpanel, docuri) => {
-		var config = getFountainConfig(vscode.Uri.parse(docuri));
-		if (change.affectsConfiguration("fountain")) {
-			if (previewpanel) {
-				var config = getFountainConfig(vscode.Uri.parse(docuri));
-				previewpanel.webview.postMessage({ command: 'updateconfig', content: config })
-			}
-		}
-	})
+    previews.forEach(p => {
+        var config = getFountainConfig(vscode.Uri.parse(p.uri));
+        if (change.affectsConfiguration("fountain"))
+            p.panel.webview.postMessage({ command: 'updateconfig', content: config })
+    });
 })
 
 vscode.window.onDidChangeTextEditorSelection(change => {
 	if(change.textEditor.document.languageId == "fountain")
 	var config = getFountainConfig(change.textEditor.document.uri);
 	if (config.synchronized_markup_and_preview) {
-		var selection = change.selections[0];
-		if (previewpanels.has(change.textEditor.document.uri.toString())) {
-			previewpanels.get(change.textEditor.document.uri.toString()).webview.postMessage({ command: 'showsourceline', content: selection.active.line, linescount: change.textEditor.document.lineCount, source: "click" });
-		}
+        var selection = change.selections[0];
+        previews.forEach(p => {
+            if(p.uri == change.textEditor.document.uri.toString())
+                p.panel.webview.postMessage({ command: 'showsourceline', content: selection.active.line, linescount: change.textEditor.document.lineCount, source: "click" });
+        });
 	}
 })
 
@@ -168,8 +204,8 @@ export class FountainPreviewSerializer implements vscode.WebviewPanelSerializer 
       // also restore any event listeners we need on it.
 
 
-      var docuri = vscode.Uri.parse(state.docuri);
-      loadWebView(docuri, webviewPanel);
+      let docuri = vscode.Uri.parse(state.docuri);
+      loadWebView(docuri, webviewPanel, state.dynamic);
       //webviewPanel.webview.postMessage({ command: 'updateTitle', content: state.title_html });
       //webviewPanel.webview.postMessage({ command: 'updateScript', content: state.screenplay_html });
     }